refactored backstack
