+++ change/@microsoft-teams-js-2ec25d83-8460-43ca-ae3b-41f2ec1d4b7e.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Refactored `backstack` initialization to resolve a `pages` bug when dynamically importing teams-js",
+  "packageName": "@microsoft/teams-js",
+  "email": "109628470+noahdarveau-MSFT@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
